EW-860 code cleanup
