+++ apps/server/src/modules/account/domain/index.ts
@@ -1,3 +1,4 @@
 export * from './account';
+export * from './account-save';
 export * from './update-account';
 export * from './update-my-account';
